Add files via upload
A retailer faces stochastic demand for a product and needs to determine the optimal reordering strategy over the next 10 days(planning horizon). At the end of the planning horizon, the new generation of the product will replace the current one, and any unsold inventory will be salvaged at a reduced price to a secondary market (e.g. an outlet store, Groupon, etc.).

Tips:
1) Edit in MATLAB the script named inventory.m
2) Modify the part of the script in lines 2-10 (provide the proper model input) 
3) Modify the part of the script in lines 18, 20, 34, 37, 41 (replace the functions poisspdf and poisscdf with binopdf and binocdf, and use the proper input for these functions)